replaced u16 with usize

+++ src/generalized_taking_game/closed_generalized/mod.rs
@@ -6,14 +6,14 @@ pub mod symmetries;
 
 #[derive(Hash, Clone, PartialEq, Eq, PartialOrd, Ord, Debug)]
 pub struct ClosedTakingGamePart {
-    sets_of_nodes: Vec<Vec<u16>>,
-    set_indices: Vec<Vec<u16>>,
-    nodes: u16,
+    sets_of_nodes: Vec<Vec<usize>>,
+    set_indices: Vec<Vec<usize>>,
+    nodes: usize,
 }
 
 //Constructors and basic functions for the GeneralizedNimGame
 impl ClosedTakingGamePart {
-    pub fn get_possible_nimbers(&self) -> Vec<u16>{
+    pub fn get_possible_nimbers(&self) -> Vec<usize>{
         if self.find_symmetry().is_some() {
             return vec![0];
         }else{
@@ -23,10 +23,10 @@ impl ClosedTakingGamePart {
     pub fn into_generalized(self) -> TakingGame {
         return TakingGame { parts: vec![self] };
     }
-    pub fn get_sets_of_nodes(&self) -> &Vec<Vec<u16>> {
+    pub fn get_sets_of_nodes(&self) -> &Vec<Vec<usize>> {
         &self.sets_of_nodes
     }
-    pub fn get_node_count(&self) -> u16 {
+    pub fn get_node_count(&self) -> usize {
         self.nodes
     }
 }
