+++ src/generalized_taking_game/closed_generalized/moves.rs
@@ -7,7 +7,7 @@ impl ClosedTakingGamePart {
     pub fn get_unique_moves(&self) -> Vec<TakingGame> {
         let mut moves = vec![];
 
-        let mut processed_sets:Vec<&Vec<u16>> = vec![];
+        let mut processed_sets:Vec<&Vec<usize>> = vec![];
 
         for group in &self.sets_of_nodes {
             processed_sets.push(group);
@@ -20,8 +20,8 @@ impl ClosedTakingGamePart {
     }
     fn append_moves_of_group(
         &self,
-        lone_nodes: Vec<u16>,
-        other_nodes: Vec<u16>,
+        lone_nodes: Vec<usize>,
+        other_nodes: Vec<usize>,
         child_games: &mut Vec<TakingGame>,
     ) {
         if other_nodes.len() > 128 {
@@ -33,19 +33,19 @@ impl ClosedTakingGamePart {
             let start_value = if lone_nodes_to_remove == 0 { 1 } else { 0 };
             for mask in start_value..mask_bound {
                 let child_game =
-                    self.get_child(&lone_nodes, &other_nodes, lone_nodes_to_remove as u16, mask);
+                    self.get_child(&lone_nodes, &other_nodes, lone_nodes_to_remove , mask);
 
                 child_games.push(child_game);
             }
         }
     }
-    fn collect_lone_nodes_and_other_nodes(&self, group: &Vec<u16>) -> (Vec<u16>, Vec<u16>) {
+    fn collect_lone_nodes_and_other_nodes(&self, group: &Vec<usize>) -> (Vec<usize>, Vec<usize>) {
         let mut lone_nodes_in_group = vec![];
         let mut other_nodes_in_group = vec![];
 
         for node in group {
             let node = *node;
-            if self.set_indices[node as usize].len() == 0 {
+            if self.set_indices[node ].len() == 0 {
                 lone_nodes_in_group.push(node);
             } else {
                 other_nodes_in_group.push(node);
@@ -56,14 +56,14 @@ impl ClosedTakingGamePart {
 
     fn get_child(
         &self,
-        lone_nodes: &Vec<u16>,
-        other_nodes: &Vec<u16>,
-        lone_nodes_to_remove: u16,
+        lone_nodes: &Vec<usize>,
+        other_nodes: &Vec<usize>,
+        lone_nodes_to_remove: usize,
         mask: u128,
     ) -> TakingGame {
         let mut nodes_to_remove = vec![];
         for i in 0..lone_nodes_to_remove {
-            nodes_to_remove.push(lone_nodes[i as usize]);
+            nodes_to_remove.push(lone_nodes[i ]);
         }
         let mut mask = mask;
         for i in 0..other_nodes.len() {
@@ -77,7 +77,7 @@ impl ClosedTakingGamePart {
         return child;
     }
     ///removes all nodes specified in the argument
-    pub fn make_move_unchecked(&self, nodes_to_remove: &mut Vec<u16>) -> TakingGame {
+    pub fn make_move_unchecked(&self, nodes_to_remove: &mut Vec<usize>) -> TakingGame {
         nodes_to_remove.sort_unstable();
 
         let mut new_groups = vec![];
